Fix bitrates over 144 kBit/s
When the desired bitrate and therefore the size of the output Superframe was
too high (over 144 kBit/s), the default output buffer size (with 2 channels:
2^11 = 2048 bytes) was exceeded. As the buffer write function applies a bitmask
to the target address, silently a rollover occured and the function continued
to (over)write at the start of the buffer. Thus the encoding output got
unusable.

To fix this, consider the maximum allowed Superframe size (for 192 kBit/s) when
the output buffer size is set.